Move apple and spotify tests into a separate folder.
